[Fix] 결과 창 이미지 적용 완료
추가 작업 사항이 있음
- 문자열 포맷팅을 위한 컴포넌트 분리
- 결과 창에서 일시정지 버튼이 눌려 일시정지 캔버스가 출력되는 현상 제거

+++ TouhouSurvival/Assets/Scripts/GameManager/GameManager.cs
@@ -301,7 +301,7 @@ namespace Unchord
             if (ShouldUpdateElapsedPlaytime)
             {
                 ElapsedPlaytime += Time.deltaTime;
-                UIManager.Instance.GameCanvas.SetTimer((int)ElapsedPlaytime);
+                UIManager.Instance.GameCanvas.SetTimer(ElapsedPlaytime);
             }
         }
